don't cancel shutdown handler after kernel.view.
When a form in a block is submitted, the form is not processed in the controller
but in the view event, because Drupal blocks are rendered by the delivery callback. If
the shutdown handler is cancelled after the controller is called, calls to
`drupal_goto()` had no effect because the shutdown handler only was active during the
call to the controller.