introduce PseudoKernel.
PseudoKernel allows Drupal's bootstrap to be treated as a service,
which means it can be injected into dependencies so that Drupal
can be only bootstrapped as needed when the request is not a Drupal
route.

Bootstrap now depends on the event dispatcher so that events can
be triggered before and after bootstrap phases.